safari-600.7-branch/LayoutTests/platform/ios-sim/svg/animations/animate-elem-05-t-drt-expected.txt

SVG 1.1 dynamic animation tests

A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rootSVGElement.getTransformToElement(path).e is -15
PASS rootSVGElement.getTransformToElement(path).f is -43
PASS rootSVGElement.getTransformToElement(path).e is -40
PASS rootSVGElement.getTransformToElement(path).f is -30
PASS rootSVGElement.getTransformToElement(path).e is -65
PASS rootSVGElement.getTransformToElement(path).f is -30
PASS rootSVGElement.getTransformToElement(path).e is -65
PASS rootSVGElement.getTransformToElement(path).f is -30
PASS successfullyParsed is true

TEST COMPLETE
